Add limited resilience against DCS outages.
This allows the primary to not demote itself but to continue online in some
cases when the DCS fails. If DCS fails, the primary queries all standbys and
compares their dcs_last_seen timestamps with its own and the timestamp of the
DCS outage. If the standbys experience the same DCS outage, continue as a
primary. If any of the standbys (i) is not reachable, (ii) has seen the DCS
since or (iii) has been promoted to primary, the (now former) primary demotes
itself as before.

This logic is in a new dcs_offline_fallback() function that is called at the
required places.