Avoid running CHECKPOINT on remote master if credentials are missing
If pg_rewind is required to be run by a standby-cluster member, we need to make
sure the main cluster leader has persisted the timeline changes to disk, so we
were running CHECKPOINT on it, which requires a remote superuser connection. If
no superuser password was configured, this would fail (#2194).

Instead, first use the rewind credentials to check if the correct timeline is
present in the main cluster leader's control file. Only if this is not the
case, run CHECKPOINT using the superuser credentials.